Computes the minimum relative effect size to achieve replication success with the sceptical p-value

The minimum relative effect size (replication to original) to achieve replication success with the sceptical p-value is computed based on the result of the original study and the corresponding variance ratio.

effectSizeReplicationSuccess(
  zo,
  c = 1,
  level = 0.025,
  alternative = c("one.sided", "two.sided"),
  type = c("golden", "nominal", "controlled")
)

Arguments

zo

Numeric vector of z-values from original studies.

c

Numeric vector of variance ratios of the original and replication effect estimates. This is usually the ratio of the sample size of the replication study to the sample size of the original study.

level

Threshold for the calibrated sceptical p-value. Default is 0.025.

alternative

Specifies if level is "one.sided" (default) or "two.sided". If "one.sided", then effect size calculations are based on a one-sided assessment of replication success in the direction of the original effect estimate.

type

Type of recalibration. Can be either "golden" (default), "nominal" (no recalibration), or "controlled". "golden" ensures that for an original study just significant at the specified level, replication success is only possible for replication effect estimates larger than the original one. "controlled" ensures exact overall Type-I error control at level level^2.

Value

The minimum relative effect size to achieve replication success with the sceptical p-value.

Details

effectSizeReplicationSuccess is the vectorized version of the internal function .effectSizeReplicationSuccess_. Vectorize is used to vectorize the function.

References

Held, L., Micheloud, C., Pawel, S. (2022). The assessment of replication success based on relative effect size. The Annals of Applied Statistics. 16:706-720. doi:10.1214/21-AOAS1502

Micheloud, C., Balabdaoui, F., Held, L. (2023). Assessing replicability with the sceptical p-value: Type-I error control and sample size planning. Statistica Neerlandica. doi:10.1111/stan.12312

See also

Author

Leonhard Held, Charlotte Micheloud, Samuel Pawel, Florian Gerber

Examples

po <- c(0.001, 0.002, 0.01, 0.02, 0.025)
zo <- p2z(po, alternative = "one.sided")

effectSizeReplicationSuccess(zo = zo, c = 1, level = 0.025,
                             alternative = "one.sided", type = "golden")
#> [1] 0.5752170 0.6338302 0.8840540 1.1347634 1.2720196

effectSizeReplicationSuccess(zo = zo, c = 10, level = 0.025,
                             alternative = "one.sided", type = "golden")
#> [1] 0.3272944 0.3792091 0.6218755 0.8837976 1.0304384
effectSizeReplicationSuccess(zo = zo, c = 10, level = 0.025,
                             alternative = "one.sided", type = "controlled")
#> [1] 0.1676487 0.1889537 0.2779059 0.3564431 0.3934514
effectSizeReplicationSuccess(zo = zo, c= 2, level = 0.025,
                             alternative = "one.sided", type = "nominal")
#> [1] 0.6869157 0.7955308 1.4460987 3.1224530       Inf

effectSizeReplicationSuccess(zo = zo, c = 2, level = 0.05,
                             alternative = "two.sided", type = "nominal")
#> [1] 0.6869157 0.7955308 1.4460987 3.1224530       Inf