Computes the minimum relative effect size to achieve significance of the replication study

The minimum relative effect size (replication to original) to achieve significance of the replication study is computed based on the result of the original study and the corresponding variance ratio.

effectSizeSignificance(
  zo,
  c = 1,
  level = 0.025,
  alternative = c("one.sided", "two.sided")
)

Arguments

zo

Numeric vector of z-values from original studies.

c

Numeric vector of variance ratios of the original and replication effect estimates. This is usually the ratio of the sample size of the replication study to the sample size of the original study.

level

Significance level. Default is 0.025.

alternative

Specifies if the significance level is "one.sided" (default) or "two.sided". If the significance level is one-sided, then effect size calculations are based on a one-sided assessment of significance in the direction of the original effect estimate.

Value

The minimum relative effect size to achieve significance in the replication study.

Details

effectSizeSignificance is the vectorized version of the internal function .effectSizeSignificance_. Vectorize is used to vectorize the function.

References

Held, L., Micheloud, C., Pawel, S. (2022). The assessment of replication success based on relative effect size. The Annals of Applied Statistics. 16:706-720. doi:10.1214/21-AOAS1502

See also

Author

Charlotte Micheloud, Samuel Pawel, Florian Gerber

Examples

po <- c(0.001, 0.002, 0.01, 0.02, 0.025)
zo <- p2z(po, alternative = "one.sided")

effectSizeSignificance(zo = zo, c = 1, level = 0.025,
                       alternative = "one.sided")
#> [1] 0.6342449 0.6809777 0.8425068 0.9543348 1.0000000

effectSizeSignificance(zo = zo, c = 1, level = 0.05,
                       alternative = "two.sided")
#> [1] 0.6342449 0.6809777 0.8425068 0.9543348 1.0000000

effectSizeSignificance(zo = zo, c = 50, level = 0.025,
                       alternative = "one.sided")
#> [1] 0.08969577 0.09630479 0.11914846 0.13496332 0.14142136