Computes the level for the sceptical p-value

The equivalent level for the sceptical p-value is computed based on the specified limiting relative effect size, the replication success level, and the alternative.

levelEquivalent(dinf, level = 0.025, alternative = c("one.sided", "two.sided"))

Arguments

dinf
The limiting relative effect size bound below which replication success is deemed to be impossible for borderline significant original studies with p-values equal to level.

level
Numeric vector of required replication success levels. Defaults to 0.025.

alternative
Either "one.sided" (default) or "two.sided". Specifies if the replication success level is one-sided or two-sided. If the replication success level is one-sided, then a one-sided level for the sceptical p-value is computed.

Value

Equivalent levels for the sceptical p-value corresponding to dmin=1. Thus, at the equivalent level, replication success is impossible for borderline significant original studies and shrinkage of the replication effect estimate.

References

Held, L. (2020). A new standard for the analysis and design of replication studies (with discussion). Journal of the Royal Statistical Society: Series A (Statistics in Society), 183, 431-448. doi: 10.1111/rssa.12493

Held, L., Micheloud, C., Pawel, S. (2021). The assessment of replication success based on relative effect size. https://arxiv.org/abs/2009.07782

See also

Author

Leonhard Held

Examples

levelEquivalent(dinf = 0.8, level = 0.025)

#> [1] 0.0332667

levelEquivalent(dinf = 0.8, level = 0.05, alternative="two.sided")

#> [1] 0.0665334