Computes the level for the sceptical p-value

The level for the sceptical p-value is computed based on the specified alternative and calibration type.

levelSceptical(
  level,
  alternative = c("one.sided", "two.sided"),
  type = c("golden", "nominal", "liberal", "controlled")
)

Arguments

level

Numeric vector of required replication success levels.

alternative

Either "one.sided" (default) or "two.sided". Specifies if the replication success level is one-sided or two-sided. If the replication success level is one-sided, then a one-sided level for the sceptical p-value is computed.

type

The calibration type can be either "golden" (default), "nominal", "liberal", or "controlled". type = "golden" ensures that for an original study just significant at the specified level, replication success is only possible if the replication effect estimate is larger than the original one. If type = "controlled", the type-I error rate is equal to level^2 (for alternative ="two.sided") or 2 \(\times\)level^2 (for alternative = "one.sided") if the variance ratio is equal to 1. The type "nominal" ensures that the type-I error rate is always smaller than level^2. Significance of both the original and replication study at level is then a necessary but not sufficient requirement for replication success. If type is "liberal" then significance of both studies is a sufficient requirement for replication success if the variance ratio is equal to 1.

Value

Levels for the sceptical p-value.

References

Held, L. (2020). A new standard for the analysis and design of replication studies (with discussion). Journal of the Royal Statistical Society: Series A (Statistics in Society), 183, 431-448. doi:10.1111/rssa.12493

Held, L. (2020). The harmonic mean chi-squared test to substantiate scientific findings. Journal of the Royal Statistical Society: Series C (Applied Statistics), 69, 697-708. doi:10.1111/rssc.12410

Held, L., Micheloud, C., Pawel, S. (2022). The assessment of replication success based on relative effect size. The Annals of Applied Statistics. 16:706-720. doi:10.1214/21-AOAS1502

Author

Leonhard Held

Examples

levelSceptical(level = 0.025, alternative = "one.sided", type = "nominal")
#> [1] 0.025
levelSceptical(level = 0.025, alternative = "one.sided", type = "liberal")
#> [1] 0.08288814
levelSceptical(level = 0.025, alternative = "one.sided", type = "controlled")
#> [1] 0.06530883
levelSceptical(level = 0.025, alternative = "one.sided", type = "golden")
#> [1] 0.06167928