Probability of replicating an effect of Killeen (2005)

Computes the probability that a replication study yields an effect estimate in the same direction as in the original study.

pReplicate(
  po = NULL,
  zo = p2z(p = po, alternative = alternative),
  c,
  alternative = "two.sided"
)

Arguments

po
Numeric vector of p-values from the original study, default is NULL.

zo
Numeric vector of z-values from the original study. Is calculated from po, if necessary.

c
The ratio of the variances of the original and replication effect estimates. This is usually the ratio of the sample size of the replication study to the sample size of the original study.

alternative
Either "two.sided" (default) or "one.sided". Specifies whether the p-value is two-sided or one-sided.

Value

The probability that a replication study yields an effect estimate in the same direction as in the original study.

Details

This extends the statistic p_rep ("the probability of replicating an effect") by Killeen (2005) to the case of possibly unequal sample sizes, see also Senn (2002).

References

Killeen, P. R. (2005). An alternative to null-hypothesis significance tests. Psychological Science, 16, 345--353. doi: 10.1111/j.0956-7976.2005.01538.x

Senn, S. (2002). Letter to the Editor, Statistics in Medicine, 21, 2437--2444.

Held, L. (2019). The assessment of intrinsic credibility and a new argument for p < 0.005. Royal Society Open Science, 6, 181534. doi: 10.1098/rsos.181534

Author

Leonhard Held

Examples

pReplicate(po = c(0.05, 0.01, 0.001), c = 1)

#> [1] 0.9171119 0.9657259 0.9900108

pReplicate(po = c(0.05, 0.01, 0.001), c = 2)

#> [1] 0.9452344 0.9822739 0.9963920

pReplicate(po = c(0.05, 0.01, 0.001), c = 2, alternative = "one.sided")

#> [1] 0.9103668 0.9712477 0.9941848

pReplicate(zo = c(2, 3, 4), c = 1)

#> [1] 0.9213504 0.9830526 0.9976611