Find multiple roots in interval

Searches the interval from lower to upper for several roots (i.e., zero's) of a univariate function f.

unirootAll(
  f,
  interval,
  lower = min(interval),
  upper = max(interval),
  n = 1000,
  tol = .Machine$double.eps^0.2,
  maxiter = 1000,
  trace = 0,
  ...
)

Arguments

f

the function for which the root is sought. f should be vectorized in the first argument.

interval

A vector containing the end-points of the interval to be searched for the root.

lower

The lower end point of the interval to be searched.

upper

The upper end point of the interval to be searched.

n

Number of subintervals on which link[stats]{uniroot} is called. Default is 1000.

tol

See help of link[stats]{uniroot}.

maxiter

See help of link[stats]{uniroot}.

trace

See help of link[stats]{uniroot}.

...

Additional named or unnamed arguments to be passed to f.

Value

A numeric vector of the roots found in the interval.

References

This function is inspired by rootSolve::uniroot.all(), package version 1.8.2.2.

See also

Author

Florian Gerber

Examples

f <- function (x) cos(2*x)^3
(roots <- unirootAll(f = f, interval = c(0, 10)))
#> [1] 0.7854345 2.3561140 3.9270804 5.4979631 7.0685198 8.6392599
f(roots) 
#> [1] -3.849102e-13 -4.165799e-12 -5.744889e-12  4.356810e-11  2.064153e-12
#> [6] -1.378717e-11