Computes the sceptical p-value and z-value

Computes sceptical p-values and z-values based on the z-values of the original and the replication study and the corresponding variance ratio. If specified, the p-values are recalibrated.

pSceptical(
  zo,
  zr,
  c,
  alternative = c("one.sided", "two.sided"),
  type = c("golden", "nominal", "liberal", "controlled")
)

zSceptical(zo, zr, c)

Arguments

zo
Numeric vector of z-values from original studies.

zr
Numeric vector of z-values from replication studies.

c
Numeric vector of variance ratios of the original and replication effect estimates. This is usually the ratio of the sample size of the replication study to the sample size of the original study.

alternative
Either "one.sided" (default) or "two.sided". If "one.sided", the sceptical p-value is based on a one-sided assessment of replication success in the direction of the original effect estimate. If "two.sided", the sceptical p-value is based on a two-sided assessment of replication success regardless of the direction of the original and replication effect estimate.

type
Type of recalibration. Can be either "golden" (default), "nominal", "liberal", or "controlled". Setting type to "nominal" corresponds to no recalibration as in Held et al. (2020). A recalibration is applied if type is "liberal", "controlled", or "golden", and the sceptical p-value can then be interpreted on the same scale as an ordinary p-value (e.g., a one-sided sceptical p-value can be thresholded at the conventional 0.025 level). See levelSceptical for details about recalibration types.

Value

pSceptical returns the sceptical p-value.

zSceptical returns the z-value of the sceptical p-value.

Details

pSceptical is the vectorized version of .pSceptical_. Vectorize is used to vectorize the function.

References

Held, L. (2020). A new standard for the analysis and design of replication studies (with discussion). Journal of the Royal Statistical Society: Series A (Statistics in Society), 183, 431-448. doi: 10.1111/rssa.12493

Held, L., Micheloud, C., Pawel, S. (2021). The assessment of replication success based on relative effect size. https://arxiv.org/abs/2009.07782

See also

Author

Leonhard Held

Examples

## no recalibration (type = "nominal") as in Held (2020)
pSceptical(zo = p2z(0.01), zr = p2z(0.02), c = 2, alternative = "one.sided",
           type = "nominal")

#> [1] 0.05790852

## recalibration with golden level as in Held, Micheloud, Pawel (2020)
pSceptical(zo = p2z(0.01), zr = p2z(0.02), c = 2, alternative = "one.sided",
           type = "golden")

#> [1] 0.02273138

## two-sided p-values 0.01 and 0.02, relative sample size 2
pSceptical(zo = p2z(0.01), zr = p2z(0.02), c = 2, alternative = "one.sided")

#> [1] 0.02273138

## reverse the studies
pSceptical(zo = p2z(0.02), zr = p2z(0.01), c = 1/2, alternative = "one.sided")

#> [1] 0.008786301

## both p-values 0.01, relative sample size 2
pSceptical(zo = p2z(0.01), zr = p2z(0.01), c = 2, alternative = "two.sided")

#> [1] 0.03496702

zSceptical(zo = 2, zr = 3, c = 2)

#> [1] 1.531634

zSceptical(zo = 3, zr = 2, c = 2)

#> [1] 1.531634